v2 picks up the SDXL coverage
